requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
psycopg2-binary==2.9.7
bcrypt==4.0.1
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1
itsdangerous==2.1.2
WTForms==3.0.1
email-validator==2.0.0
gunicorn==21.2.0